Moved code out of init
